+++ ByteBusters/Game/MazeGenerator/Source Files/placeTetrisBlocks.cpp
@@ -32,8 +32,8 @@ const std::vector<std::vector<int>> tetrisShapes[] = { //!< A blokk alakok kiala
   */
 void placeTetrisBlocks(std::vector<std::vector<int>>& grid, float targetDensity) {
     //!< A méret lekérése
-    int numRows = grid.size();
-    int numCols = grid[0].size();
+    size_t numRows = grid.size();
+    size_t numCols = grid[0].size();
 
     //seedeles a veletlenszeruseghez
     //std::srand(static_cast<unsigned int>(std::time(nullptr)));
